Register Based VM (partial) and better regcode

+++ vm/read.c
@@ -13,7 +13,7 @@ uint32_t read_u32(uint8_t *code, int idx) {
 uint64_t read_u64(uint8_t *code, int idx) {
     union {
         uint8_t *restrict u8;
-        uint32_t *restrict u64;
+        uint64_t *restrict u64;
     } conv = {code + idx};
     return *conv.u64;
 }
